backend/requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
python-dotenv==1.0.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1
cryptography==41.0.7

# Database & Caching
psycopg2-binary==2.9.9
django-redis==5.4.0
redis==5.0.1

# Task Queue
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.0

# API & Documentation
django-filter==23.3
drf-spectacular==0.26.5

# GraphQL
graphene-django==3.2.0
django-graphql-jwt==0.4.0

# Internationalization
django-modeltranslation==0.18.11

# File Handling & Static Files
whitenoise==6.6.0
Pillow==10.1.0

# Data Processing & ML
pandas==2.1.1
numpy==1.26.0
scikit-learn==1.3.1
sqlalchemy==2.0.22
openpyxl==3.1.2

# Monitoring & Logging
django-extensions==3.2.3
django-debug-toolbar==4.2.0
sentry-sdk==1.38.0

# Development & Testing
pytest==7.4.2
pytest-django==4.7.0
black==23.9.1
flake8==6.1.0
coverage==7.3.2
factory-boy==3.3.0

# Production
gunicorn==21.2.0
django-health-check==3.17.0

# Cloud & Storage
boto3==1.34.0
django-storages==1.14.2

# Security & Compliance
django-ratelimit==4.1.0
django-guardian==2.4.0
django-two-factor-auth==1.15.0
qrcode==7.4.2

# Monitoring & Metrics
prometheus-client==0.19.0
django-prometheus==2.3.1

# Advanced Features
django-import-export==3.3.1
django-activity-stream==2.0.0
django-notifications-hq==1.8.3
django-rest-swagger==2.2.0

# SSO & OAuth
social-auth-app-django==5.4.0
python-saml==1.15.0

# Data Processing
great-expectations==0.18.5
# ydata-profiling==4.6.4  # Commented out - not compatible with Python 3.13

# Utilities
requests==2.31.0
python-dateutil==2.8.2
validators==0.22.0
faker==20.1.0
factory-boy==3.3.0
